Refactor method post_message via messages
- update docs about Slack, Hipchat
